refactor(backend): move crawler routes into package
