+++ docs/architecture/1-6_后端架构改造方案.md
@@ -180,7 +180,7 @@ Agno Agent 是执行引擎，新架构的 planner 通过它来调用工具完成
 | `llm/auto_todo_detection_service.py` | 自动待办检测 | 同上 |
 | `llm/ocr_todo_extractor.py` | OCR 待办提取 | 同上 |
 | `routers/automation.py` + `services/automation_task_service.py` | 自动化任务 | 简单定时任务 |
-| `routers/crawler.py` + `services/plugin_manager.py` | 爬虫插件 | MediaCrawler 集成 |
+| `routers/crawler/` + `services/plugin_manager.py` | 爬虫插件 | MediaCrawler 集成 |
 | `routers/floating_capture.py` | 悬浮窗截图提取 | 轻量工具 |
 | `routers/preview.py` | 文件预览 | 轻量工具 |
 | `routers/ocr.py` | 手动 OCR | 已被后台 job 替代 |
