+++ server/routers/crawler/runtime.py
@@ -16,7 +16,7 @@ from typing import Any, TextIO
 
 from fastapi import APIRouter, HTTPException
 
-from routers.crawler_common import (
+from .common import (
     extract_config_value,
     get_crawler_dir,
     get_sign_srv_dir,
